It may happen that a directory is deleted while being archived.
Try to not throw in the finaly clause when this happens, so the primary crash is
not masked.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/203030043

+++ isolateserver.py
@@ -1984,7 +1984,7 @@ def archive_files_to_storage(storage, files, blacklist):
     _uploaded_files = storage.upload_items(items_to_upload)
     return results
   finally:
-    if tempdir:
+    if tempdir and os.path.isdir(tempdir):
       file_path.rmtree(tempdir)
